# py/dj/free2z/openapi/f2z.yaml
openapi: 3.0.3
info:
  title: f2z
  version: 0.1.0
  description: Free2z is a tool for content and peer-2-peer giving that does not compromise
    on privacy
paths:
...
  /api/creator/:
    get:
      operationId: creator_list
      description: List Creators on Free2z
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - creator
      security:
      - cookieAuth: []
      - basicAuth: []
      - knoxApiToken: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCreatorListList'
          description: ''
  /api/creator/{username__iexact}/:
    get:
      operationId: creator_retrieve
      description: List Creators on Free2z
      parameters:
      - in: path
        name: username__iexact
        schema:
          type: string
        required: true
      tags:
      - creator
      security:
      - cookieAuth: []
      - basicAuth: []
      - knoxApiToken: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatorDetail'
          description: ''
...
components:
  schemas:
...
    AccessEnum:
      enum:
      - private
      - members
      - public
      type: string
      description: |-
        * `private` - private
        * `members` - members
        * `public` - public
...
    CreatorDetail:
      type: object
      description: Creator detail serializer
      properties:
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        total:
          type: string
          format: decimal
          pattern: ^-?\d{0,2}(?:\.\d{0,8})?$
        is_verified:
          type: boolean
          description: Once a Creator is verified, all of their pages are
        zpages:
          type: integer
          readOnly: true
        avatar_image:
          $ref: '#/components/schemas/FeaturedImage'
        banner_image:
          $ref: '#/components/schemas/FeaturedImage'
        full_name:
          type: string
          maxLength: 128
        description:
          type: string
          maxLength: 1024
        p2paddr:
          type: string
          maxLength: 255
        can_stream:
          type: boolean
          description: Returns True if the creator can start a livestream of the given
            type.
          readOnly: true
        member_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,6}(?:\.\d{0,0})?$
          nullable: true
          description: Price for 30 days of membership, livestream access, etc
      required:
      - avatar_image
      - banner_image
      - can_stream
      - username
      - zpages
    CreatorList:
      type: object
      description: Listing page serializer for Creators
      properties:
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        total:
          type: string
          format: decimal
          pattern: ^-?\d{0,2}(?:\.\d{0,8})?$
        is_verified:
          type: boolean
          description: Once a Creator is verified, all of their pages are
        zpages:
          type: integer
          readOnly: true
        avatar_image:
          $ref: '#/components/schemas/FeaturedImage'
        banner_image:
          $ref: '#/components/schemas/FeaturedImage'
        full_name:
          type: string
          maxLength: 128
        p2paddr:
          type: string
          maxLength: 255
        member_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,6}(?:\.\d{0,0})?$
          nullable: true
          description: Price for 30 days of membership, livestream access, etc
      required:
      - avatar_image
      - banner_image
      - username
      - zpages
...
    FeaturedImage:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        url:
          type: string
          readOnly: true
        name:
          type: string
          readOnly: true
        thumbnail:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        title:
          type: string
          maxLength: 128
        description:
          type: string
          maxLength: 255
        access:
          $ref: '#/components/schemas/AccessEnum'
        mime_type:
          $ref: '#/components/schemas/MimeTypeEnum'
      required:
      - created_at
      - id
      - mime_type
      - name
      - thumbnail
      - updated_at
      - url
...
    MimeTypeEnum:
      enum:
      - image/jpeg
      - image/png
      - image/gif
      - image/svg+xml
      - image/webp
      - video/mp4
      type: string
      description: |-
        * `image/jpeg` - image/jpeg
        * `image/png` - image/png
        * `image/gif` - image/gif
        * `image/svg+xml` - image/svg+xml
        * `image/webp` - image/webp
        * `video/mp4` - video/mp4
...
    PaginatedCreatorListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/CreatorList'
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    knoxApiToken:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"